ポリスマクロ定義修正

+++ HappyHomeDesigner/HappyHomeDesigner/load_csv.cpp
@@ -88,10 +88,10 @@ void LoadCsv(void)
 			furniture->scl.z = (float)atof(tkn[LOAD_CSV_SCL_Z]);
 
 			furniture->type = atoi(tkn[LOAD_CSV_TYPE]);
-			furniture->house_num = atof(tkn[LOAD_CSV_HOUSE_NUM]);
+			furniture->house_num = atoi(tkn[LOAD_CSV_HOUSE_NUM]);
 			furniture->weight = (float)atof(tkn[LOAD_CSV_WEIGHT]);
-			furniture->price = atof(tkn[LOAD_CSV_PRICE]);
-			furniture->ratio = atof(tkn[LOAD_CSV_RATIO]);
+			furniture->price = atoi(tkn[LOAD_CSV_PRICE]);
+			furniture->ratio = atoi(tkn[LOAD_CSV_RATIO]);
 			
 			furniture++;	// Ÿ‚Ìƒf[ƒ^‚É
 		}
